Move types to its own pkg
